order place bug fix

+++ studentPantry/src/main/java/com/student/pantry/studentPantry/service/OrderHistoryService.java
@@ -2,16 +2,21 @@ package com.student.pantry.studentPantry.service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.student.pantry.studentPantry.dto.OrderHistoryDto;
 import com.student.pantry.studentPantry.entity.OrderHistory;
+import com.student.pantry.studentPantry.entity.Products;
 import com.student.pantry.studentPantry.repository.OrderHistoryRepository;
+import com.student.pantry.studentPantry.repository.ProductJpa;
 import com.student.pantry.studentPantry.repository.ShoppingCartRepository;
 import com.student.pantry.studentPantry.response.OrderHistoryResponse;
 
@@ -24,14 +29,16 @@ public class OrderHistoryService {
    private final EmailService emailService;
     
     private final UserServiceImpl userServiceImpl;
+    private final ProductJpa productJpa;
 
 
     @Autowired
-    public OrderHistoryService(OrderHistoryRepository orderHistoryRepository, ShoppingCartRepository shoppingCartRepository,UserServiceImpl userServiceImpl,EmailService emailService) {
+    public OrderHistoryService(OrderHistoryRepository orderHistoryRepository, ShoppingCartRepository shoppingCartRepository,UserServiceImpl userServiceImpl,EmailService emailService,ProductJpa productJpa) {
         this.orderHistoryRepository = orderHistoryRepository;
         this.shoppingCartRepository=shoppingCartRepository;
         this.emailService=emailService;
         this.userServiceImpl=userServiceImpl;
+        this.productJpa=productJpa;
     }
 
     public OrderHistoryResponse createOrder(OrderHistoryDto orderHistoryDto) {
@@ -49,6 +56,12 @@ public class OrderHistoryService {
 
         // Convert DTO to Entity and save to the database
         OrderHistory orderHistoryEntity = convertDtoToEntity(orderHistoryDto);
+        boolean updateQuantity= updateProductQuantity(orderHistoryEntity.getItems());
+        if(!updateQuantity){
+             response=new OrderHistoryResponse("Some product in the order are invalid!!", orderPlaced);
+             return response;
+
+        }
         OrderHistory savedOrder = orderHistoryRepository.save(orderHistoryEntity);
         orderPlaced=convertEntityToDto(savedOrder);
         shoppingCartRepository.deleteByUserID(savedOrder.getUserId());
@@ -57,6 +70,7 @@ public class OrderHistoryService {
                  if(userEmail!=null && savedOrder.getOrderId()!=null) {
                 	 emailService.sendOrderConfirmation(userEmail, savedOrder.getOrderId());
                  }
+
         }
         else{
             response=new OrderHistoryResponse("Cannot place order!! Number of items exceeded 15", orderPlaced);
@@ -97,6 +111,42 @@ public class OrderHistoryService {
         }
         return true;
     }
+
+    @Transactional
+    private boolean updateProductQuantity(Map<Long, Integer> items) {
+        try{
+
+            List<Products> products=new ArrayList<>();
+        for (Map.Entry<Long, Integer> entry : items.entrySet()) {
+            Long productId = entry.getKey();
+            Integer newQuantity = entry.getValue();
+            Optional<Products> product=productJpa.findById(productId);
+            System.out.println("prduct+" + product.toString());
+            if(product.isPresent()){
+                int oldQuantity=product.get().getProductQuantity();
+                Products newProduct=product.get();
+                newProduct.setProductQuantity(oldQuantity-newQuantity);
+                System.out.println("prduct+" + newProduct.toString());
+                products.add(newProduct);
+                 System.out.println("prduct list+" + products);
+                //productJpa.save(product.get());
+            }
+            else{
+                System.err.println("Product does not exist");
+                return false;
+            }
+            
+        }
+        productJpa.saveAll(products);
+        
+        return true; 
+    }
+    catch(Exception e){
+        System.out.println("Exception + "+e.getLocalizedMessage());
+        return false;
+    }
+    }
+    
     public List<OrderHistory> findOrdersByUserIdAndMonth(Long userId, int month) {
         LocalDateTime start = LocalDateTime.of(LocalDate.now().getYear(), month, 1, 0, 0);
         LocalDateTime end = start.plusMonths(1);
